+++ src/mifdrv.h
@@ -33,9 +33,9 @@ void mif_drawpixel(const Canvas *canvas, const VPoint *vp);
 void mif_drawpolyline(const Canvas *canvas, const VPoint *vps, int n, int mode);
 void mif_fillpolygon(const Canvas *canvas, const VPoint *vps, int nc);
 void mif_drawarc(const Canvas *canvas,
-    const VPoint *vp1, const VPoint *vp2, int a1, int a2);
+    const VPoint *vp1, const VPoint *vp2, double a1, double a2);
 void mif_fillarc(const Canvas *canvas,
-    const VPoint *vp1, const VPoint *vp2, int a1, int a2, int mode);
+    const VPoint *vp1, const VPoint *vp2, double a1, double a2, int mode);
 void mif_putpixmap(const Canvas *canvas,
     const VPoint *vp, int width, int height, char *databits,
     int pixmap_bpp, int bitmap_pad, int pixmap_type);
