+++ src/psdrv.h
@@ -66,9 +66,9 @@ void ps_drawpixel(const Canvas *canvas, const VPoint *vp);
 void ps_drawpolyline(const Canvas *canvas, const VPoint *vps, int n, int mode);
 void ps_fillpolygon(const Canvas *canvas, const VPoint *vps, int nc);
 void ps_drawarc(const Canvas *canvas,
-    const VPoint *vp1, const VPoint *vp2, int a1, int a2);
+    const VPoint *vp1, const VPoint *vp2, double a1, double a2);
 void ps_fillarc(const Canvas *canvas,
-    const VPoint *vp1, const VPoint *vp2, int a1, int a2, int mode);
+    const VPoint *vp1, const VPoint *vp2, double a1, double a2, int mode);
 void ps_putpixmap(const Canvas *canvas,
     const VPoint *vp, int width, int height, char *databits,
     int pixmap_bpp, int bitmap_pad, int pixmap_type);
